@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   cupertino_icons: ^0.1.3
@@ -26,8 +26,8 @@ dependencies:
     sdk: flutter
   flutter_cache_manager_hive:
     path: ../
-  hive: ^1.4.1+1
-  hive_flutter: ^0.3.0+2
+  hive: ^2.0.4
+  hive_flutter: ^1.1.0
 
 dev_dependencies:
   flutter_test:
