@@ -4,19 +4,19 @@ version: 0.0.9
 homepage: https://github.com/vishna/flutter_cache_manager_hive
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   clock: ^1.0.1
   file: ^6.1.0
   flutter:
     sdk: flutter
-  flutter_cache_manager: ^2.1.0
-  hive: ^1.4.1+1
-  http: "^0.12.0+2"
+  flutter_cache_manager: ^3.1.2
+  hive: ^2.0.4
+  http: ^0.13.0
   meta: ^1.1.8
-  path: "^1.6.4"
-  path_provider: "^1.4.0"
-  pedantic: "^1.8.0+1"
-  rxdart: '>=0.23.1 <0.25.0'
-  uuid:  "^2.0.2"
+  path: ^1.6.4
+  path_provider: ^2.0.0
+  pedantic: ^1.8.0+1
+  rxdart: '>=0.26.0 <0.28.0'
+  uuid:  ^3.0.4
